some bug fix for slam6d and fbr
git-svn-id: https://svn.code.sf.net/p/slam6d/code/trunk@840 51f1e6e7-7289-49ad-8748-f09e43fc6daa

+++ src/slam6d/fbr/registration.cc
@@ -13,7 +13,7 @@ using namespace std;
 
 namespace fbr{
 
-  registration::registration(unsigned int minD, double minE, unsigned int minI, registration_method method){
+  registration::registration(double minD, double minE, unsigned int minI, registration_method method){
     minDistance = minD;
     minError = minE;
     minInlier = minI;
@@ -182,7 +182,7 @@ namespace fbr{
       }
   }
 
-  unsigned int registration::getMinDistance(){
+  double registration::getMinDistance(){
     return minDistance;
   }
